Update action to return update and dismiss methods
