+++ packages/core/src/notify.ts
@@ -175,48 +175,3 @@ export function setCustomNotificationProps(
 
   return customizedNotification
 }
-
-export function addCustNotification(updatedNotification: CustomNotification): {
-  dismiss: () => void
-  update: UpdateNotification
-} {
-  const customNotificationError =
-    validateCustomNotification(updatedNotification)
-
-  if (customNotificationError) {
-    throw customNotificationError
-  }
-
-  const notification = setCustomNotificationProps(updatedNotification)
-  addCustomNotification(notification)
-
-  const dismiss = () => removeNotification(notification.id)
-
-  function update(notificationUpdate: CustomNotification): {
-    dismiss: () => void
-    update: UpdateNotification
-  } {
-    const customNotificationError =
-      validateCustomNotification(updatedNotification)
-
-    if (customNotificationError) {
-      throw customNotificationError
-    }
-    notificationUpdate.id = notification.id
-    addCustomNotification(notificationUpdate)
-
-
-    return {
-      dismiss,
-      update
-    }
-  }
-
-  addCustomNotification(notification)
-
-
-  return {
-    dismiss,
-    update
-  }
-}
